Added UIRectangle and converted UISurface to a base class

+++ Bolt-Engine/Bolt-Core/src/Core/Image.cpp
@@ -69,7 +69,7 @@ namespace Bolt
 		result.Width = width;
 		result.Height = height;
 		result.Components = Components;
-		result.Pixels = new byte[width * height * Components];
+		result.Pixels = new byte[(int64_t)width * (int64_t)height * (int64_t)Components];
 
 		for (int i = 0; i < result.Width; i++)
 		{
